requirements.txt
google-genai>=1.0.0
pymongo[srv]>=4.6.0
python-dotenv>=1.0.0
numpy>=1.26.0
pydantic>=2.0.0
fastapi>=0.115.0
python-multipart>=0.0.9
groq>=0.9.0
mangum>=0.19.0
google-search-results>=2.4.0
apify-client>=1.8.0
